+++ src/adapter/graphql/src/mutations/flows_mut/dataset_flow_runs_mut.rs
@@ -20,6 +20,7 @@ use super::{
     FlowNotScheduled,
 };
 use crate::prelude::*;
+use crate::queries::Flow;
 use crate::{utils, LoggedInGuard};
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -49,10 +50,13 @@ impl DatasetFlowRunsMut {
 
         ensure_scheduling_permission(ctx, &self.dataset_handle).await?;
 
+        // TODO: for some datasets launching manually might not be an option:
+        //   i.e., root datasets with push sources require input data to arrive
+
         let flow_service = from_catalog::<dyn fs::FlowService>(ctx).unwrap();
         let logged_account = utils::get_logged_account(ctx);
 
-        let res = flow_service
+        let flow_state = flow_service
             .trigger_manual_flow(
                 Utc::now(),
                 fs::FlowKeyDataset::new(self.dataset_handle.id.clone(), dataset_flow_type.into())
@@ -66,7 +70,7 @@ impl DatasetFlowRunsMut {
             })?;
 
         Ok(TriggerFlowResult::Success(TriggerFlowSuccess {
-            flow: res.into(),
+            flow: Flow::new(flow_state),
         }))
     }
 
@@ -93,7 +97,7 @@ impl DatasetFlowRunsMut {
         match res {
             Ok(flow_state) => Ok(CancelScheduledTasksResult::Success(
                 CancelScheduledTasksSuccess {
-                    flow: flow_state.into(),
+                    flow: Flow::new(flow_state),
                 },
             )),
             Err(e) => match e {