+++ src/adapter/graphql/src/queries/flows/mod.rs
@@ -7,15 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-/////////////////////////////////////////////////////////////////////////////////////////
+mod flow;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FlowStatus {
-    Waiting,
-    Queued,
-    Scheduled,
-    Running,
-    Finished,
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////
+pub(crate) use flow::*;