+++ src/adapter/graphql/tests/tests/test_gql_dataset_flow_runs.rs
@@ -11,18 +11,27 @@
 
 use std::sync::Arc;
 
-use async_graphql::{value, Value};
+use async_graphql::value;
 use chrono::Utc;
+use container_runtime::ContainerRuntime;
 use dill::Component;
 use event_bus::EventBus;
 use indoc::indoc;
 use kamu::testing::{MetadataFactory, MockDependencyGraphRepository};
-use kamu::{DatasetRepositoryLocalFs, DependencyGraphServiceInMemory};
+use kamu::{
+    DataFormatRegistryImpl,
+    DatasetRepositoryLocalFs,
+    DependencyGraphServiceInMemory,
+    EngineProvisionerNull,
+    ObjectStoreRegistryImpl,
+    PollingIngestServiceImpl,
+};
 use kamu_core::{
     auth,
     CreateDatasetResult,
     DatasetRepository,
     DependencyGraphRepository,
+    PollingIngestService,
     SystemTimeSourceDefault,
 };
 use kamu_flow_system::{FlowID, FlowServiceRunConfig, FlowServiceTestDriver};
@@ -32,9 +41,9 @@ use kamu_flow_system_inmem::{
     FlowEventStoreInMem,
     FlowServiceInMemory,
 };
-use kamu_task_system::{TaskEventFinished, TaskID, TaskOutcome};
+use kamu_task_system::{TaskEventFinished, TaskEventRunning, TaskID, TaskOutcome};
 use kamu_task_system_inmem::{TaskSchedulerInMemory, TaskSystemEventStoreInMemory};
-use opendatafabric::{DatasetID, DatasetKind};
+use opendatafabric::{DatasetID, DatasetKind, FAKE_ACCOUNT_ID};
 
 use crate::utils::{authentication_catalogs, expect_anonymous_access_error};
 
@@ -71,7 +80,67 @@ async fn test_trigger_ingest_root_dataset() {
                                     "__typename": "Flow",
                                     "flowId": "0",
                                     "status": "QUEUED",
-                                    "outcome": Value::Null
+                                    "outcome": null
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
+
+    let request_code = FlowRunsHarness::list_flows_query(&create_result.dataset_handle.id);
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{:?}", response);
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetPollingIngest",
+                                            "datasetId": create_result.dataset_handle.id.to_string(),
+                                            "ingestedRecordsCount": null,
+                                        },
+                                        "status": "QUEUED",
+                                        "outcome": null,
+                                        "timing": {
+                                            "runningSince": null,
+                                            "finishedAt": null,
+                                        },
+                                        "tasks": [],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
                                 }
                             }
                         }
@@ -118,7 +187,67 @@ async fn test_trigger_execute_query_derived_dataset() {
                                     "__typename": "Flow",
                                     "flowId": "0",
                                     "status": "QUEUED",
-                                    "outcome": Value::Null
+                                    "outcome": null
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        })
+    );
+
+    let request_code = FlowRunsHarness::list_flows_query(&create_derived_result.dataset_handle.id);
+    let response = schema
+        .execute(
+            async_graphql::Request::new(request_code.clone())
+                .data(harness.catalog_authorized.clone()),
+        )
+        .await;
+
+    assert!(response.is_ok(), "{:?}", response);
+    assert_eq!(
+        response.data,
+        value!({
+            "datasets": {
+                "byId": {
+                    "flows": {
+                        "runs": {
+                            "listFlows": {
+                                "nodes": [
+                                    {
+                                        "flowId": "0",
+                                        "description": {
+                                            "__typename": "FlowDescriptionDatasetExecuteQuery",
+                                            "datasetId": create_derived_result.dataset_handle.id.to_string(),
+                                            "transformedRecordsCount": null,
+                                        },
+                                        "status": "QUEUED",
+                                        "outcome": null,
+                                        "timing": {
+                                            "runningSince": null,
+                                            "finishedAt": null,
+                                        },
+                                        "tasks": [],
+                                        "initiator": {
+                                            "id": FAKE_ACCOUNT_ID,
+                                            "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                        },
+                                        "primaryTrigger": {
+                                            "__typename": "FlowTriggerManual",
+                                            "initiator": {
+                                                "id": FAKE_ACCOUNT_ID,
+                                                "accountName": auth::DEFAULT_ACCOUNT_NAME,
+                                            }
+                                        },
+                                        "startCondition": null,
+                                    }
+                                ],
+                                "pageInfo": {
+                                    "hasPreviousPage": false,
+                                    "hasNextPage": false,
+                                    "currentPage": 0,
+                                    "totalPages": 1,
                                 }
                             }
                         }
@@ -580,6 +709,7 @@ async fn test_cancel_already_succeeded_flow() {
     let flow_id = FlowRunsHarness::extract_flow_id_from_trigger_response(&response_json);
 
     let flow_task_id = harness.mimic_flow_scheduled(flow_id).await;
+    harness.mimic_task_running(flow_task_id).await;
     harness.mimic_task_completed(flow_task_id).await;
 
     let mutation_code =
@@ -669,6 +799,10 @@ impl FlowRunsHarness {
 
     fn new_custom(dependency_graph_repo: MockDependencyGraphRepository) -> Self {
         let tempdir = tempfile::tempdir().unwrap();
+        let run_info_dir = tempdir.path().join("run");
+        let cache_dir = tempdir.path().join("cache");
+        std::fs::create_dir(&run_info_dir).unwrap();
+        std::fs::create_dir(&cache_dir).unwrap();
 
         let catalog_base = dill::CatalogBuilder::new()
             .add::<EventBus>()
@@ -690,6 +824,16 @@ impl FlowRunsHarness {
             .add_value(FlowServiceRunConfig::new(chrono::Duration::seconds(1)))
             .add::<TaskSchedulerInMemory>()
             .add::<TaskSystemEventStoreInMemory>()
+            .add::<DataFormatRegistryImpl>()
+            .add_value(ContainerRuntime::default())
+            .add_builder(
+                PollingIngestServiceImpl::builder()
+                    .with_cache_dir(cache_dir)
+                    .with_run_info_dir(run_info_dir),
+            )
+            .bind::<dyn PollingIngestService, PollingIngestServiceImpl>()
+            .add::<EngineProvisionerNull>()
+            .add::<ObjectStoreRegistryImpl>()
             .build();
 
         // Init dataset with no sources
@@ -715,6 +859,7 @@ impl FlowRunsHarness {
                 MetadataFactory::dataset_snapshot()
                     .kind(DatasetKind::Root)
                     .name("foo")
+                    .push_event(MetadataFactory::set_polling_source().build())
                     .build(),
             )
             .await
@@ -748,6 +893,23 @@ impl FlowRunsHarness {
             .unwrap()
     }
 
+    async fn mimic_task_running(&self, task_id: TaskID) {
+        let flow_service_test_driver = self
+            .catalog_authorized
+            .get_one::<dyn FlowServiceTestDriver>()
+            .unwrap();
+        flow_service_test_driver.mimic_running_started();
+
+        let event_bus = self.catalog_authorized.get_one::<EventBus>().unwrap();
+        event_bus
+            .dispatch_event(TaskEventRunning {
+                event_time: Utc::now(),
+                task_id,
+            })
+            .await
+            .unwrap();
+    }
+
     async fn mimic_task_completed(&self, task_id: TaskID) {
         let flow_service_test_driver = self
             .catalog_authorized
@@ -770,6 +932,95 @@ impl FlowRunsHarness {
         response_json["datasets"]["byId"]["flows"]["runs"]["triggerFlow"]["flow"]["flowId"]
             .as_str()
             .unwrap()
+    }
+
+    fn list_flows_query(id: &DatasetID) -> String {
+        indoc!(
+            r#"
+            {
+                datasets {
+                    byId (datasetId: "<id>") {
+                        flows {
+                            runs {
+                                listFlows {
+                                    nodes {
+                                        flowId
+                                        description {
+                                            __typename
+                                            ... on FlowDescriptionDatasetCompaction {
+                                                datasetId
+                                                originalBlocksCount
+                                                resultingBlocksCount
+                                            }
+                                            ... on FlowDescriptionDatasetExecuteQuery {
+                                                datasetId
+                                                transformedRecordsCount
+                                            }
+                                            ... on FlowDescriptionDatasetPollingIngest {
+                                                datasetId
+                                                ingestedRecordsCount
+                                            }
+                                            ... on FlowDescriptionDatasetPushIngest {
+                                                datasetId
+                                                sourceName
+                                                inputRecordsCount
+                                                ingestedRecordsCount
+                                            }
+                                        }
+                                        status
+                                        outcome
+                                        timing {
+                                            runningSince
+                                            finishedAt
+                                        }
+                                        tasks {
+                                            taskId
+                                            status
+                                            outcome
+                                        }
+                                        initiator {
+                                            id
+                                            accountName
+                                        }
+                                        primaryTrigger {
+                                            __typename
+                                            ... on FlowTriggerInputDatasetFlow {
+                                                datasetId
+                                                flowType
+                                                flowId
+                                            }
+                                            ... on FlowTriggerManual {
+                                                initiator {
+                                                    id
+                                                    accountName
+                                                }
+                                            }
+                                        }
+                                        startCondition {
+                                            __typename
+                                            ... on FlowStartConditionBatching {
+                                                thresholdNewRecords
+                                            }
+                                            ... on FlowStartConditionThrottling {
+                                                intervalSec
+                                            }
+                                        }
+                                    }
+                                    pageInfo {
+                                        hasPreviousPage
+                                        hasNextPage
+                                        currentPage
+                                        totalPages
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            "#
+        )
+        .replace("<id>", &id.to_string())
     }
 
     fn trigger_flow_mutation(id: &DatasetID, dataset_flow_type: &str) -> String {