+++ src/domain/flow-system/src/entities/flow/flow_state.rs
@@ -21,14 +21,26 @@ pub struct FlowState {
     pub flow_id: FlowID,
     /// Flow key
     pub flow_key: FlowKey,
-    /// Activating at time
-    pub activate_at: Option<DateTime<Utc>>,
+    /// Primary flow trigger
+    pub primary_trigger: FlowTrigger,
+    /// Start condition (if defined)
+    pub start_condition: Option<FlowStartCondition>,
+    /// Timing records
+    pub timing: FlowTimingRecords,
     /// Associated task IDs
     pub task_ids: Vec<TaskID>,
     /// Flow outcome
     pub outcome: Option<FlowOutcome>,
-    /// Termiation time (cancel or abort)
-    pub terminated_at: Option<DateTime<Utc>>,
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct FlowTimingRecords {
+    /// Activating at time
+    pub activate_at: Option<DateTime<Utc>>,
+    /// Started running at time
+    pub running_since: Option<DateTime<Utc>>,
+    /// Finish time (success or cancel/abort)
+    pub finished_at: Option<DateTime<Utc>>,
 }
 
 impl FlowState {
@@ -36,12 +48,14 @@ impl FlowState {
     pub fn status(&self) -> FlowStatus {
         if self.outcome.is_some() {
             FlowStatus::Finished
+        } else if self.timing.running_since.is_some() {
+            FlowStatus::Running
         } else if !self.task_ids.is_empty() {
             FlowStatus::Scheduled
-        } else if self.activate_at.is_some() {
+        } else if self.timing.activate_at.is_some() {
             FlowStatus::Queued
         } else {
-            FlowStatus::Draft
+            FlowStatus::Waiting
         }
     }
 }
@@ -59,14 +73,19 @@ impl Projection for FlowState {
                     event_time: _,
                     flow_id,
                     flow_key,
-                    trigger: _,
+                    trigger,
                 }) => Ok(Self {
                     flow_id,
                     flow_key,
-                    activate_at: None,
+                    primary_trigger: trigger,
+                    start_condition: None,
+                    timing: FlowTimingRecords {
+                        activate_at: None,
+                        running_since: None,
+                        finished_at: None,
+                    },
                     task_ids: vec![],
                     outcome: None,
-                    terminated_at: None,
                 }),
                 _ => Err(ProjectionError::new(None, event)),
             },
@@ -78,12 +97,15 @@ impl Projection for FlowState {
                     E::StartConditionDefined(FlowEventStartConditionDefined {
                         event_time: _,
                         flow_id: _,
-                        start_condition: _,
+                        start_condition,
                     }) => {
                         if s.outcome.is_some() || !s.task_ids.is_empty() {
                             Err(ProjectionError::new(Some(s), event))
                         } else {
-                            Ok(s)
+                            Ok(FlowState {
+                                start_condition: Some(*start_condition),
+                                ..s
+                            })
                         }
                     }
                     E::Queued(FlowEventQueued {
@@ -95,7 +117,10 @@ impl Projection for FlowState {
                             Err(ProjectionError::new(Some(s), event))
                         } else {
                             Ok(FlowState {
-                                activate_at: Some(*activate_at),
+                                timing: FlowTimingRecords {
+                                    activate_at: Some(*activate_at),
+                                    ..s.timing
+                                },
                                 ..s
                             })
                         }
@@ -116,7 +141,7 @@ impl Projection for FlowState {
                         flow_id: _,
                         task_id,
                     }) => {
-                        if s.outcome.is_some() || s.activate_at.is_none() {
+                        if s.outcome.is_some() || s.timing.activate_at.is_none() {
                             Err(ProjectionError::new(Some(s), event))
                         } else {
                             let mut task_ids = s.task_ids.clone();
@@ -125,13 +150,33 @@ impl Projection for FlowState {
                             Ok(FlowState { task_ids, ..s })
                         }
                     }
+                    E::TaskRunning(FlowEventTaskRunning {
+                        event_time,
+                        flow_id: _,
+                        task_id,
+                    }) => {
+                        if s.outcome.is_some()
+                            || s.timing.activate_at.is_none()
+                            || !s.task_ids.contains(task_id)
+                        {
+                            Err(ProjectionError::new(Some(s), event))
+                        } else {
+                            Ok(FlowState {
+                                timing: FlowTimingRecords {
+                                    running_since: Some(*event_time),
+                                    ..s.timing
+                                },
+                                ..s
+                            })
+                        }
+                    }
                     E::TaskFinished(FlowEventTaskFinished {
                         event_time,
                         flow_id: _,
                         task_id,
                         task_outcome,
                     }) => {
-                        if !s.task_ids.contains(task_id) {
+                        if !s.task_ids.contains(task_id) || s.timing.running_since.is_none() {
                             Err(ProjectionError::new(Some(s), event))
                         } else if s.outcome.is_some() {
                             // Ignore for idempotence motivation
@@ -140,11 +185,18 @@ impl Projection for FlowState {
                             match task_outcome {
                                 TaskOutcome::Success => Ok(FlowState {
                                     outcome: Some(FlowOutcome::Success),
+                                    timing: FlowTimingRecords {
+                                        finished_at: Some(*event_time),
+                                        ..s.timing
+                                    },
                                     ..s
                                 }),
                                 TaskOutcome::Cancelled => Ok(FlowState {
                                     outcome: Some(FlowOutcome::Cancelled),
-                                    terminated_at: Some(event_time.clone()),
+                                    timing: FlowTimingRecords {
+                                        finished_at: Some(*event_time),
+                                        ..s.timing
+                                    },
                                     ..s
                                 }),
                                 // TODO: support retries
@@ -169,7 +221,10 @@ impl Projection for FlowState {
                         } else {
                             Ok(FlowState {
                                 outcome: Some(FlowOutcome::Aborted),
-                                terminated_at: Some(event_time.clone()),
+                                timing: FlowTimingRecords {
+                                    finished_at: Some(*event_time),
+                                    ..s.timing
+                                },
                                 ..s
                             })
                         }