+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Initial support for local predefined Admin users
+- New GraphQL APIs to list summary information on dataset flows
 
 ## [0.151.0] - 2024-01-09
 ### Added